// cargodelivery-front/generated/models/FuelStationIn.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Cargo delivery
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { Point } from './Point';
import {
    PointFromJSON,
    PointFromJSONTyped,
    PointToJSON,
} from './Point';

/**
 * 
 * @export
 * @interface FuelStationIn
 */
export interface FuelStationIn {
    /**
     * 
     * @type {string}
     * @memberof FuelStationIn
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof FuelStationIn
     */
    description: string;
    /**
     * 
     * @type {Point}
     * @memberof FuelStationIn
     */
    coordinates: Point;
}

/**
 * Check if a given object implements the FuelStationIn interface.
 */
export function instanceOfFuelStationIn(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "name" in value;
    isInstance = isInstance && "description" in value;
    isInstance = isInstance && "coordinates" in value;

    return isInstance;
}

export function FuelStationInFromJSON(json: any): FuelStationIn {
    return FuelStationInFromJSONTyped(json, false);
}

export function FuelStationInFromJSONTyped(json: any, ignoreDiscriminator: boolean): FuelStationIn {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'name': json['name'],
        'description': json['description'],
        'coordinates': PointFromJSON(json['coordinates']),
    };
}

export function FuelStationInToJSON(value?: FuelStationIn | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'name': value.name,
        'description': value.description,
        'coordinates': PointToJSON(value.coordinates),
    };
}